Minor change to formatting
I imagined that the link was intended for the word "launch", so I added the necessary brackets.

On a side note, I don't quite understand why that particular part of the angular github page was linked to (https://github.com/angular/angular.js/blob/master/CHANGELOG.md#100-temporal-domination-2012-06-13)? If that has a particular meaning, it's at least not very obvious to me.