Added id to each course
In exercise 2.4, the "course object" becomes an element of an array of course objects and we are asked to edit our application in a way that can handle an arbitrary number of courses.
In order to populate React's "key" attribute, we need a unique identifier that sets apart each course from the rest of the courses.
The provided array of courses lacks such a property. We could use the array's index, but that may not be a best practice to learn for a student.
The lack of a property suitable for the "key" attribute may be the intent of the author to make us wonder how to solve this problem, which is a good thing. Or it may be a slip.
In any case, I edited the exercise to include such an identifier.
Hope that helps.