Unsupported routes handler fix
The text says that the unsupported routes handler should be the last middleware, while it needs to be the next to the last. This is pointed out further down in the text, but it's more explicit and understandable to point that out right away, especially since a previous code snippet shows that.