Changed expected HTML response in test
When trying the test we get an error saying it expected 201 and received 200. I tried to see if I had any error in my code and even downloaded the current version from the course, it was still not working despite the course saying it should. Later on in the course, the code is changed from 201 to 200 and it then works. I think it was intended to be 200 since the beginning so it can work properly which is why I am submitting this change