Uniform names of action creator functions
The action creator for new notes is called 'createNote' through all of the document, except in two instances where it is called 'noteCreation'. This can generate confusion.

Also, arrow functions parameters in action creators are always wrapped into round brackets through all of the document, except for an instance of 'content' which is not. Uniformed also that to the document standard.