The ellipsis wrongly implies it should change
Neither the variable Course that's passed into Header prop nor Header itself changes. The code on this line remains identical to what went before. Using the ellipsis suggests to students that something should be different about this line when no difference is necessary.

This part of the code only changes in the next section, 1.5, when the structure of the constants change.