Update part6a.md
In Dan Abramov's Medium article about presentational and container components which is linked in this part in the 'More Components' section, he has an updated edit about not splitting components such way. "This text is left intact for historical reasons but don’t take it too seriously" is on the update. It gives a sense that this section isn't updated.