part1a: visually fix the link in the footer
The "greeintg app created by" link is on a new line in the source code, which results in no whitespace rendering between "by" and "mluukkai". I'm sure there's a way to force a whitespace (perhaps using nbsp?) but I'm a react noob and I just noticed that if I bring the link to the same line as the text, the whitespace is rendered.