part1a: ask learners to also remove setupTests.js
This file is not being used and is not discussed anywhere in this chapter. All other files in the directory are asked to be deleted except for this one, and I don't see any reason why this one should remain if it isn't mentioned or used - it just adds confusion.