Part 2c: Another way to secure API key
It seems that Windows users can have trouble with the environment variables. I tried many different things from StackOverflow, but nothing seemed to work. I think it might be a good idea to add another way of working with the API keys, so that the problems with environment variables do not cause students to avoid the exercise itself. I think saving all sensitive information in a separate config file and not committing that file is a good practice. But I also understand if you feel like environment variables are the way to go.