Modified description of exercise 6.19.
Since the students have been using redux with the newer state hooks in part 6, the description of exercise 6.19 had the incorrect assumption that the redux store was being passed through props and that the students would have to install the 'react-redux' package. Very minimal changes have been made to update the exercise description so students would not be confused.