Update part4c.md
I might be wrong but there is a problem here.

The current code saves user to the db, but since we need to add notes to user, we only need to update user. The current version should give an error of "username appears" cause we "save" it again.

I propose update. As I could see, I updated the parts where it needs.