Update part3c.md
In practice, use of the toJSON method doesn't seem mandatory to get JSON formatted data, but the course does mention it and the examples don't include it, I didn't understand why. Please either rewrite the relevant instructions or include the method in the examples!

This problem causes discrepancies between code examples in part 3c and part 3d, where we can see toJSON used in snippets written in previous parts of the course.