Remove NoteForm highlight (and className?) in 5c
The NoteForm code adds a "formDiv" class to the div surrounding the form, and that change is highlighted, but this class isn't actually necessary for the NoteForm test. The test grabs the form itself and the input field. 

I don't think the change is significant enough to be highlighted, since it's not relevant to test, and maybe adding the class isn't necessary as well.