Update part1b.md
The term used should be arguments and not parameters. Function parameters are the names listed in the function's definition. Function arguments are the real values passed to the function. Parameters are initialized to the values of the arguments supplied.

Please refer this - https://developer.mozilla.org/en-US/docs/Glossary/Parameter