matching the store code in Redux DevTools with 6-2 branch
There is `export default store` at the end of the code snippet, but we have not refactored the store into its own `store.js` file at this point yet. The store definition is still in `index.js` in the [6-2 branch](https://github.com/FullStack-HY/redux-notes/blob/part6-2/src/index.js) (which is mentioned shortly after the snippet.)

On that note, is Exercise 6.9 a bit too early? Creating `store.js` is done in the next section Part6c. I'm fine with doing it before the lesson does it, it's simple enough.