Grammar and alt text fixes for part 0b
I added alt texts for most of the images in this part, most of which are somewhat generic. As of now, screen readers just read "full stack content" for every image in the part, likely the entire course if this is representative, which is kind of lackluster. I'd appreciate it if some attention could be paid to this; I myself am a screen reader user and can only guess at what an image is actually displaying from context.