Note on connection strings and troubleshooting
Added possible troubleshoot for issues with requests made to MongoDb database. This solution was found at https://stackoverflow.com/questions/55499175/how-to-fix-error-querysrv-erefused-when-connecting-to-mongodb-atlas . I am contributing it because this small problem caused me a couple of days of scrutinizing my code for a problem that my code did not have. If the details as to why this works can be added to my contribution, that would be appreciated.