Update part7b.md
restcountries.com offer a v3 as well that doesn't work with the base app that is provided. Maybe it's a part of the course to figure that out but since the exercises focuses on using hooks it felt a bit farfetched. Suggesting a link that goes to v2 and a clarification in the link.