exerc 7.7, correct API doc url
Current URL takes to the same page as https://restcountries.com/  -  the front page of the documentation. Documentation starts with version 3. It seems that the JSON data format returned from 'full name' endpoint differs from v2. And I assume that the ready coded part of exerc 7.7 expects data returned as v2. I refactured the code to make it work and later discovered that I used v3 and the exercise expects v2. I found this out by comparing the FI and EN course materials.