About HTTP Request Types: typo/grammar 1103
"This means that if a request has side-effects, then the result should be *the* same regardless of how many times the request is sent."

Is "same" always preceded with "the"?

Yes, "same" is always preceded with "the". This is not always the case in speech e.g. "Same thing here". But this is actually a case of elision of the word "the".
In writing, however, "the" always precedes "same" i.e. "it is same thing" is incorrect, and should be "It is the same thing."

The word same is usually used with the definite article. However, it can be used with any central determiner which marks the noun phrase as definite:

these same ideas
those very same people
my same friend
whose same idea
Ben's same problems