minor grammar corrections in part 2c
Explanations for some of the changes:

On line 43 "the" isn't necessary, without it the sentence can be more succinct. 
On line 57 "saving *them* too" is more readable.
Missing comma in line 57 after "application".
On line 92 punctuation must follow N.B. 
On line 128 "i.e." was missing a comma. The recommended punctuation for 'i.e.' in these cases is 'i.e.,', e.g.[,] this Cambridge Dictionary example: "The hotel is closed during the off season, i.e., from October to March."
The following article explains it quite nicely: https://jakubmarian.com/comma-after-i-e-and-e-g/

On line 248 the "starting a new one, otherwise there will be trouble:" sentence is a run-on. It needs a semicolon to separate the independent clauses. 

On line 302 the correct grammar for "completed" in this situation would be "been completed".

On line 332 a comma is missing before and. An Oxford ("serial") comma should be used when writing lists.

On line 337 the correct grammar is "on" and not "in"