minor grammar changes to part2d
Explanations for some of the changes:

On line 174, "E.g," is missing a comma at the end. The recommended punctuation for 'E.g.' in these cases is 'E.g.,', i.e.[,] this Cambridge Dictionary example: "The hotel is closed during the off season, i.e., from October to March."
The following article explains it quite nicely: https://jakubmarian.com/comma-after-i-e-and-e-g/

On line 222, there's a spelling mistake; the past tense of spread is spread. "Spreaded" would be considered a spelling error. On the same line, there should be a comma after "In practice", since it's used here to emphasize a rhetorical shift.

On line 236, "in new object" was missing a "the".

On line 487, there's a missing comma before "and". An Oxford ("serial") comma should be used when writing lists.