Add more context about which project we're on
This change reminds readers that we are adding code to the notes project, not the phonebook app. Since the exercises from the previous section (2b) deal with the phonebook app, it can be easy to forget to switch back to the notes project without any context in the first paragraphs. The only cue currently is the `notes` property in the json file which is easy to miss, and later down in the effects-hook section. I've added a reference to the notes project to avoid confusion and mistakes.