Would advise updating pictures
Some of the pictures used on. this page are from when this project was on Heroku. The project now exists on the helsinki.fi website and the rest of the content reflects that. It is a minor change, but could assist those who could be confused about the content vs the screenshots.