Fix console warning
As the code is structure now, a console warning is thrown:
```
You are calling ReactDOMClient.createRoot() on a container that has already been passed to createRoot() before. Instead, call root.render() on the existing root instead if you want to update it.
```
This PR refactors the code so that we are not calling createRoot() multiple times.