Create part3c.md
without providing more info on the set method on the schema it may seem to a beginer like myself that we are setting some sort of instance method, i did some research and found out that this is not the case and that we are actually configuring options on the schema. i have added some links to the docs that explain what we are actually doing here. this seems like quite and advanced use of mongoose and i wonder if it is only adding confusion to make this change so early on.