Update part3b.md
In line 98, about the configuration of port for Fly.io or Heroku, I tried the port number without double quotes but it did not work for me, as I tried couple of times. But when I placed the port number inside double quotes then it started working. Is there anything that I have misunderstood or does the port number should be inside double quotes?