Update part4d.md
I think it seems a bit minor issue, but I suggest a small change in  the [part where we parse the authorization header](https://fullstackopen.com/en/part4/token_authentication#limiting-creating-new-notes-to-logged-in-users). It is in line 156.

A few reasons for this suggestions are the followings:
- We don't rely on a specific index(here `7`) for parsing, using regex instead
- According to [RFC6750 Section 2.1](https://www.rfc-editor.org/rfc/rfc6750#section-2.1) they specify that the Authorization header starts with "Bearer". 
(I think there are some issues regarding this "B" case sensitivity(for example, this issue: https://github.com/golang/oauth2/issues/113), but I think it would be better to stick to what's conventional at the moment)
- Moreover, the current code `authorization.toLowerCase().startsWith('bearer ')` allows some undesirable cases like `beARER' or `BEARER', let alone `bearer`, due to the method `toLowerCase()`

Please let me know if there is something I am misunderstanding or something else that is wrong. 
Thank you for fantastic resources for learning webdev!