fix: link on markdown
When clicking the link from the full stack open website, (section 6.19. -6.21) the link directs me to https://fullstackopen.com/en/part6/(https://github.com/fullstack-hy2020/query-anecdotes).
I believe this is happening because in the markdown, the link is wrapped in double parentheses [this project]((linkhere)).

I have removed the double parentheses to resolve this issue.