Update part4b.md
Currently, line 755 fails the tests. Comparing resultNote.body to noteToView fails since resultNote.body.date is a date formatted as a String whereas noteToView.date is a Date object.

The solution is to simply use the date String in resultNote.body.date to create a new Date object and reassign to itself.