Change wording in part4b.md
Replace two instances of the phrase "create to the tests directory a file" with "add to the tests directory a file".

While "create to the directory" can be understood, it is an uncommon (and potentially incorrect) wording. Instead, the phrase "add to the directory" makes more sense. Alternatively, "create in the directory" would also make sense.