Update part-1d, missing syntax
The 3 Button components in the example were directly calling the setToValue() function instead of wrapping it into a function and then calling it (just like in the screenshot below). When copy and pasted into vscode, React would error out due to an infinite loop. This fix matches the correct way like in the screenshot.