Update part7b.md
Removed link to [old React Router page on hooks](https://reacttraining.com/react-router/web/api/Hooks) because the content no longer exists. The [new docs](https://reactrouter.com/en/main) do not have a route to an overview of hooks because documentation for each hook is directly accessible via the sidebar. Since the link to the new docs are already embedded earlier in the sentence, this link is no longer necessary.