Import, not create  context
Before we put the context creation in the './src/contexts/AuthStorageContext.js' file. It makes sense to import this here and use it. Otherwise, it will throw an error since the context created in the useAuthStorage is not provided in the 'App.js'. 

(If it makes more sense, to create the context here, we should add explanation on why and delete the 'contexts' directory).