Update Fetching a single resource code
The id parameter from the request object is received as a string. So, when we run the array.find() method and do the '===' (equal value and equal type) it will return undefined because the id and the note.id do not match in 'type' (note.id is a number). 

This fix is to do '==' (equal to) and the array.find() will return the wanted value.