Update part3a.md
Problem Statement 
When we first see the piece of code Math.max(...notes.map(n => n.id)), it is a bit intriguing and I had to look for other resources to see how it works.  The explanation to the line of code was at the end of the notes.

Solution 

To move the explanation right after the piece of code was first introduced.