Edited a minor thing in part4a.md
Through the whole course, text was formal and directed to a group of people, the people write the content.

This one felt a little out of the context. I know that you, @mluukkai have done a lot to this course and I love your way of writing. I am not sure if you wrote it, but even I respect that.

If this change isn't welcomed, please don't take it personal. I am here a student willing to help, 😇