Update part1d.md
The word "Handles" is capitalized and it stands after the word "Button" which is inside the <i> tags. This makes they seem like there is no space between them ("ButtonHandles").

It might create confusion about the naming of the component, when sentence should be like this:  "Button handles the functionality of each feedback submission button".