Update part1d.md
Use a comma before additional information.

I was also wondering: The title "Utilization of Large language models" seems to appear in the github repo. However, in the reading material, this title reads "Utilization of language language models". Is it just the case that the material has not been updated since that was fixed in the repo?