Update part4c.md
The test of creating new note that has a userID in it has been giving me error. After debugging it I found that we were updating the user document on itself and saving it by "user.save() "
I changed the way we update the user to "await User.findByIdAndUpdate() " and it worked.