revert changing localhost:3001 address
The problem trying to fix is that while initially the course material creates and uses the endpoint 'http://localhost:3001/api/testing/reset', when it refers to using baseUrl it uses as an example the route 'http://localhost:5173/api/tests/reset', which is never used in the course material.
As the course first mentions to "replace all the addresses in the tests from http://localhost:3001/api/... to http://localhost:5173/api/...", it makes sense to use as an example the route 'http://localhost:5173/api/testing/reset'. For this reason, I suggest to change only /tests/ to /testing/ as this might lead to confusion.