Update part9c.md
Fix the code from the part 'typing the request response', the res.send is duplicated and also the import { Response } from 'express' is missing, I don't know if it's intended but I'm dumb so I thought that I just missed something and decided to re-read everything from the start.  
I'm not sure if this is the correct way but if it's not please change it so people like me can save time. Thanks.