Update part6a.md
Remember that I said the quality of the course has been steadily declining after part 2? Here is a typical example. As someone who is learning it for the first time, it is reasonable, if not a common-sense to make clear to the new learner WHERE the code goes. App.jsx? main.jsx? or a seperate file then later to be imported? An increasing number of mistakes like this in the later chapters suggest that as time went on, the author just let the course go.