Update part6d.md
Using object instead of an array, you won't need any of the helper functions or any of the adjustments in other files. Furthermore, you avoid digging yourself into a hole(as in a possible future mistake) when you use object keys instead of array indexing. I hope you can see what I am trying to do here. I was indeed surprised when I saw you used array. In addition, the helper functions, even though I know what they do, in the future when someone else try to understand this code, it would require whomever read this code to go around an additional circle to understand the code, instead of the directness of not using helper functions. So maybe helper functions are not always as helpful, sometimes they can be counter-helpful and make the code more confusing for when another person reads them.