Update part2d.md
This might not be the best solution but, I got stuck not understand why my error was not being caught. It was because I already put the catches into the Notes.jsx service script, consequently the catch in App.jsx did not catch anything and I had an array of notes with an undefined item.