Add clarification about Vite exported variables
While doing this part I was getting frustrated with my requests not going through to the weather provider. Ultimately tracking that down to my key not being registered. I believe it would be best to place a clarification that the VITE_ prefix is not optional but required. Placement of this clarification can be revised.