Update part9c.md
The original code gives an error because you set the return type to be "DiaryEntry[]" but the res actually returns values from "getNonSensitiveEntries()" which has a "NonSensitiveDiaryEntry" type(without the "comment" property"). It gave an error - 'comment' is declared here. The fix is to set the "NonSensitiveDiaryEntry[]" type instead of "DiaryEntry[]" type.