Correct information about env context availability
To the best of my knowledge, the removed information was not correct. I have tried myself to access environment variable with `github.env`, and it doesn't work. I then went online only to find advice such as this one, which worked for me and which I changed in the materials:
https://stackoverflow.com/questions/73558652/github-actions-how-to-use-environment-variable-at-job-level