Update part0b.md
The sentence "The URL new_note_spa, that new notes are sent to, does not adhere to current best practices." is either lacking a comma or has a comma too much, since the "that new notes are sent to" modifies "The URL". 
Also, replacing the "that" with a "which" or "where" enhances readability.