Add GitHub link to example applications reference
The course material mentions that the code for the example applications can be found on GitHub, but no link was provided.  
I couldn’t find a direct repo or folder that contains all of the example applications in one place, but I added a link to the official course repository (https://github.com/fullstack-hy2020/fullstack-hy2020.github.io) so that readers have at least one starting point to navigate from.