Clarify HTTP GET route handling in exercise 3.18
The text "handling of the api/persons/:id and info routes" is a bit ambiguous since exercise 3.17 requires you to also implement a api/persons/:id route, as did exercise 3.15. I think adding HTTP GET in the exercise makes it a bit more clear to the student that it is the GET /api/persons/:id and GET /info routes that are to be updated in this exercise.