Update explanation of React state updates
Clarify timing of state updates in React. I was not sure based on the text if calling this function multiple times is a mistake, because if the update could happen before the current component function is finished, then it's a concurrency issue. (The word asynchronous lead me to think that.) Checking the react docs it seems to be the case that this never happens, the updates are done in between the 2 component runs.